+++ R/Learner.R
@@ -235,9 +235,8 @@ Learner = R6Class("Learner",
       }
 
       train_row_ids = if (!is.null(row_ids)) row_ids else task$row_roles$use
-      test_row_ids = task$row_roles$test
 
-      learner_train(learner, task, train_row_ids = train_row_ids, test_row_ids = test_row_ids, mode = mode)
+      learner_train(learner, task, train_row_ids = train_row_ids, mode = mode)
 
       # store data prototype
       proto = task$data(rows = integer())